+++ packages/yoga/src/Dialog/web/Footer.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Box from '../../Box';
 
 const Footer = props => (
-  <Box as="footer" d="flex" gap="large" ml="auto" {...props} />
+  <Box as="footer" d="flex" justifyContent="center" gap="large" {...props} />
 );
 
 Footer.displayName = 'Dialog.Footer';