+++ packages/yoga/src/Header/Header.theme.js
@@ -1,0 +1,12 @@
+const Header = ({ spacing }) => ({
+  padding: {
+    xxs: spacing.medium,
+    lg: spacing.huge,
+  },
+  height: {
+    xxs: 56,
+    lg: 72,
+  },
+});
+
+export default Header;