+++ packages/yoga/src/Input/web/Input.jsx
@@ -109,10 +109,12 @@ const Input = React.forwardRef(
           full={full}
           label={label}
           style={style}
+          value={value}
         >
           <Field
             {...props}
             {...{
+              label,
               cleanable,
               disabled,
               error,
@@ -179,6 +181,7 @@ Input.propTypes = {
   onClean: func,
   /** when max length helper is unnecessary to appear */
   hideMaxLength: bool,
+  placeholder: string,
 };
 
 Input.defaultProps = {
@@ -196,6 +199,7 @@ Input.defaultProps = {
   onChange: () => {},
   onClean: () => {},
   hideMaxLength: false,
+  placeholder: undefined,
 };
 
 export default Input;