+++ packages/yoga/src/index.js
@@ -28,6 +28,10 @@ import Avatar from './Avatar';
 import BottomSheet from './BottomSheet';
 import Dialog from './Dialog';
 import Divider from './Divider';
+import Header from './Header';
+import Banner from './Banner';
+import Heading from './Heading';
+import Feedback from './Feedback';
 
 export {
   ThemeProvider,
@@ -64,4 +68,8 @@ export {
   BottomSheet,
   Dialog,
   Divider,
+  Header,
+  Banner,
+  Heading,
+  Feedback,
 };