+++ packages/helpers/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/Gympass/yoga/compare/@gympass/yoga-helpers@1.0.2...@gympass/yoga-helpers@1.0.3) (2022-06-09)
+
+**Note:** Version bump only for package @gympass/yoga-helpers
+
+
+
+
+
+## [1.0.2](https://github.com/Gympass/yoga/compare/@gympass/yoga-helpers@1.0.1...@gympass/yoga-helpers@1.0.2) (2022-06-07)
+
+**Note:** Version bump only for package @gympass/yoga-helpers
+
+
+
+
+
 ## [1.0.1](https://github.com/Gympass/yoga/compare/@gympass/yoga-helpers@1.0.0...@gympass/yoga-helpers@1.0.1) (2021-09-24)
 
 **Note:** Version bump only for package @gympass/yoga-helpers